修正long long 型数据校验时长度问题导致的校验值不匹配

+++ Classes/comm/Util/DataManager.cpp
@@ -649,7 +649,7 @@ bool bailin::util::DataManagerUtil::SetSecurityData(const char *pszKey, long lon
         *pllData += llAddData;
         
         // 回填校验值
-        unsigned long nCheckCode = bailin::util::crypto::Crc32(pllData, sizeof(long));
+        unsigned long nCheckCode = bailin::util::crypto::Crc32(pllData, sizeof(long long));
         // m_pDictionarySecurityData->setObject(CCString::createWithFormat("%lu", nCheckCode), pszKey);
         m_mapSecurityData[string(pszKey)] = nCheckCode;
     }
